+++ app/Views/admin/leave/index.php
@@ -3,35 +3,6 @@
 
 <div class="p-4 sm:ml-64 mt-14">
 
-    <?php if (isset($_SESSION['success'])): ?>
-        <div id="alert-success" class="flex items-center p-4 mb-4 text-green-800 rounded-base bg-neutral-primary-soft" role="alert">
-            <svg class="flex-shrink-0 w-4 h-4" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" fill="currentColor" viewBox="0 0 20 20">
-                <path d="M10 .5a9.5 9.5 0 1 0 9.5 9.5A9.51 9.51 0 0 0 10 .5ZM9.5 4a1.5 1.5 0 1 1 0 3 1.5 1.5 0 0 1 0-3ZM12 15H8a1 1 0 0 1 0-2h1v-3H8a1 1 0 0 1 0-2h2a1 1 0 0 1 1 1v4h1a1 1 0 0 1 0 2Z" />
-            </svg>
-            <span class="sr-only">Info</span>
-            <div class="ms-3 text-sm font-medium">
-                <?= htmlspecialchars($_SESSION['success']) ?>
-            </div>
-        </div>
-        <?php unset($_SESSION['success']); ?>
-    <?php endif; ?>
-
-    <?php if (isset($_SESSION['error'])): ?>
-        <div id="alert-danger" class="flex items-center p-4 mb-4 text-red-800 rounded-base bg-neutral-primary-soft" role="alert">
-            <svg class="flex-shrink-0 w-4 h-4" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" fill="currentColor" viewBox="0 0 20 20">
-                <path d="M10 .5a9.5 9.5 0 1 0 9.5 9.5A9.51 9.51 0 0 0 10 .5ZM9.5 4a1.5 1.5 0 1 1 0 3 1.5 1.5 0 0 1 0-3ZM12 15H8a1 1 0 0 1 0-2h1v-3H8a1 1 0 0 1 0-2h2a1 1 0 0 1 1 1v4h1a1 1 0 0 1 0 2Z" />
-            </svg>
-            <span class="sr-only">Danger</span>
-            <div class="ms-3 text-sm font-medium">
-                <?= htmlspecialchars($_SESSION['error']) ?>
-            </div>
-        </div>
-        <?php unset($_SESSION['error']); ?>
-    <?php endif; ?>
-
-
-
-
     <!-- Statistik pengajuan cuti  -->
     <?php if (isset($statistics) && !empty($statistics)): ?>
         <div class="bg-neutral-primary-soft shadow-xs rounded-base p-6 mb-6 border border-default">